refactor and add children
