+++ src/components/Lists.tsx
@@ -1,7 +1,49 @@
-import { isString, uniqueId } from "lodash";
+import { isString, merge, uniqueId } from "lodash";
 import Link from "next/link";
 
 import { Page } from "contentlayer/generated";
+
+export function buildTree(
+  paths: string[],
+  allPages: Page[],
+  filter?: (value: string) => boolean
+): Record<string, Page> | string {
+  let tree = {};
+  if (filter) {
+    paths = paths.filter(filter);
+  }
+
+  paths.forEach((path) => {
+    tree = merge(tree, buildTreeInt(allPages, path, path));
+  });
+
+  return tree;
+}
+
+export function getPaths(allPages: Page[]): string[] {
+  const paths: string[] = allPages.map((page) => page.path);
+  return paths.sort();
+}
+
+export function buildTreeInt(
+  allPages: Page[],
+  path: string,
+  fullPath: string
+): Record<string, Page> | string {
+  if (!path.includes("/")) {
+    const data = allPages.find((page) => page.path === fullPath);
+    if (!data) {
+      throw new Error(`Could not find page for ${path}`);
+    }
+
+    return { [path]: data };
+  }
+
+  const pieces = path.split("/");
+  return {
+    [pieces[0]]: buildTreeInt(allPages, pieces.slice(1).join("/"), fullPath),
+  } as unknown as Record<string, Page> | string;
+}
 
 function Tree({ items }: { items: string | Record<string, Page> }) {
   return (
